moved template functions
